fix 32bit tick time overflow issue.

+++ ffrdpdemuxer/ffrdpc.c
@@ -63,7 +63,7 @@ static void* ffrdpc_thread_proc(void *argv)
         }
 
         if (!(ffrdpc->status & TS_CONNECTED)) {
-            if (get_tick_count() - ticktryconnect > 1000) {
+            if ((int32_t)get_tick_count() - (int32_t)ticktryconnect > 1000) {
                 ffrdp_send(ffrdpc->ffrdp, "1", 1);
                 ticktryconnect = get_tick_count();
             }
@@ -90,7 +90,7 @@ static void* ffrdpc_thread_proc(void *argv)
         }
 
         ffrdp_update(ffrdpc->ffrdp);
-        if ((ffrdpc->status & TS_CONNECTED) && get_tick_count() - tickgetframe > 3000) {
+        if ((ffrdpc->status & TS_CONNECTED) && (int32_t)get_tick_count() - (int32_t)tickgetframe > 3000) {
             printf("server lost !\n");
             ffrdp_free(ffrdpc->ffrdp); ffrdpc->ffrdp = NULL;
             ffrdpc->head = ffrdpc->tail = ffrdpc->size = 0;
